moved structs to types.go
