+++ core/types.go
@@ -1,0 +1,26 @@
+package core
+
+import (
+	"github.com/aws/aws-sdk-go/service/support"
+	"github.com/aws/aws-sdk-go/service/support/supportiface"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+// SupportClientImpl ...
+type SupportClientImpl struct {
+	SupportClient supportiface.SupportAPI
+}
+
+// SupportClient ...
+type SupportClient interface {
+	RequestServiceLimitsRefreshLoop()
+	DescribeServiceLimitsCheckResult() (*support.TrustedAdvisorCheckResult, error)
+}
+
+// SupportExporter ...
+type SupportExporter struct {
+	region        string
+	supportClient SupportClient
+	metricsUsed   map[string]*prometheus.Desc
+	metricsLimit  map[string]*prometheus.Desc
+}
